refactor: Move SNS publish logic to controller
